Fixed flash probing on reset + simplified the trampoline
In reset-init, I am directly probing the flash,
which will initialize Flash when needed for flashing.
That way, I don't need to initialize flash on every auto-probe.
The simplified trampoline is not necessary, but to other side, it is good,
because it does not modify RA (and called function should save it
and then restore if needed)

Change-Id: Ic2f3ba40d3862947bd574bc1fb197f1655e98026